requirements.txt
beautifulsoup4
requests
pycairo
